backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data processing
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1

# Date parsing
python-dateutil==2.8.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0

# Code quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# CORS
python-dotenv==1.0.0

# Vector embeddings for semantic field matching
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
torch==2.1.0  # Required by sentence-transformers
chromadb==0.4.22  # Vector database for efficient similarity search

# AI Services
google-generativeai==0.3.2  # Google Gemini AI for intelligent file analysis
